Fix minor adjustments and refinements
